Implement a SyncmerIterator
This allows to iterate over all the syncmers in a sequence with very little
memory. Unlike make_string_to_hashvalues_open_syncmers_canonical, no
string_hashes and pos_to_seq_coordinate vectors are returned (which can be
quite large depending on contig size). Instead, the syncmers are returned
one by one.
